tests/pyproject.toml
[project]
name = "cell-type-annotation-agent"
version = "0.1.0"
description = "Production-style LangGraph agent for single-cell cell type annotation from marker genes"
authors = [{name = "OpenAI Demo Scaffold"}]
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "pydantic>=2.7.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "python-dotenv>=1.0.1",
    "rapidfuzz>=3.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
